feat(llm/llc): deposit flow performances improvements
