+++ libs/ledger-live-common/src/deposit/index.ts
@@ -1,4 +1,3 @@
-import { useMappedAssets } from "./useMappedAssets.hook";
 import { useGroupedCurrenciesByProvider } from "./useGroupedCurrenciesByProvider.hook";
 
-export { useMappedAssets, useGroupedCurrenciesByProvider };
+export { useGroupedCurrenciesByProvider };
