+++ libs/ledger-live-common/src/deposit/mock.ts
@@ -16,21 +16,6 @@ export const MOCK = [
   },
   {
     $type: "Token",
-    contract: "0xfd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9",
-    data: {
-      img: "https://assets.coingecko.com/coins/images/325/large/Tether.png?1668148663",
-      marketCapRank: 3,
-    },
-    ledgerId: "arbitrum/erc20/tether_usd",
-    name: "Tether USD",
-    network: "arbitrum",
-    providerId: "tether",
-    reason: null,
-    status: "Ok",
-    ticker: "USDT",
-  },
-  {
-    $type: "Token",
     contract: "0xdac17f958d2ee523a2206206994597c13d831ec7",
     data: {
       img: "https://assets.coingecko.com/coins/images/325/large/Tether.png?1668148663",
@@ -39,36 +24,6 @@ export const MOCK = [
     ledgerId: "ethereum/erc20/usd_tether__erc20_",
     name: "Tether USD",
     network: "ethereum",
-    providerId: "tether",
-    reason: null,
-    status: "Ok",
-    ticker: "USDT",
-  },
-  {
-    $type: "Token",
-    contract: "0x94b008aa00579c1307b0ef2c499ad98a8ce58e58",
-    data: {
-      img: "https://assets.coingecko.com/coins/images/325/large/Tether.png?1668148663",
-      marketCapRank: 3,
-    },
-    ledgerId: "optimism/erc20/tether_usd",
-    name: "Tether USD",
-    network: "optimism",
-    providerId: "tether",
-    reason: null,
-    status: "Ok",
-    ticker: "USDT",
-  },
-  {
-    $type: "Token",
-    contract: "0x66e428c3f67a68878562e79a0234c1f83c208770",
-    data: {
-      img: "https://assets.coingecko.com/coins/images/325/large/Tether.png?1668148663",
-      marketCapRank: 3,
-    },
-    ledgerId: "cronos/erc20/tether_usd",
-    name: "Tether USD",
-    network: "cronos",
     providerId: "tether",
     reason: null,
     status: "Ok",
